requirements/yaml.txt
yamllint==1.35.1
PyYAML==6.0.1
ruamel.yaml==0.18.6
